feat(url-ingestion): enhance pipeline with reachability checks and requeue support
Add URL reachability verification before processing, implement intelligent
requeue mechanism for unprocessed items, and improve content processing
pipeline with better error handling and formatting. Includes comprehensive
test coverage for new functionality.

- Add reachability checks for URL validation
- Implement requeue-unprocessed script and pipeline integration
- Enhance raw processing with improved categorization
- Improve content formatting and cleanup
- Add test coverage for new modules
- Update configuration and pipeline prompts

+++ stubs/litellm.pyi
@@ -24,7 +24,7 @@ def completion(
     api_base: str | None,
     temperature: float,
     max_tokens: int,
-    timeout: int | None = None,
+    timeout: float | None = None,
     stream: bool | None = None,
     response_format: Mapping[str, object] | None = None,
 ) -> ModelResponse: ...
